pyproject.toml
[project]
name = "ebook-manager"
version = "0.1.0"
description = "Lightweight ebook manager with Icecream UI and AI summaries"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[exts]>=2.0.23",
    "aiosqlite>=0.19.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "jinja2>=3.1.2",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "google-genai>=0.3.0",
    "groq>=0.4.0",
    "openai>=1.0.0",
    "httpx>=0.25.2",
    "pillow>=10.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pypdf>=3.17.0",
    "pymupdf>=1.24.0",
    "pymobi>=0.1.3",
    "gtts>=2.5.0",
    "edge-tts>=6.1.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
pythonpath = ["backend"]
